Renamed cli files
