+++ src/cruceGameCurses/main.c
@@ -1,4 +1,4 @@
-#include "output.h"
+#include "cli.h"
 #include <curses.h>
 #include <locale.h>
